+++ R/tt_paginate.R
@@ -237,7 +237,6 @@ setMethod(
            max_width = NULL,
            fontspec = NULL,
            col_gap = 3) {
-
     new_dev <- open_font_dev(fontspec)
     if (new_dev) {
       on.exit(close_font_dev())