+++ NEWS.md
@@ -1,4 +1,8 @@
-## rtables 0.6.15.9003
+## rtables 0.6.16.9001
+
+### New Features
+ * Added `restrict_facets` function factory for use with `make_split_fun`
+ * Exportd previously internal `make_subset_expr` for use when constructing custom splitting behavior
 
 ## rtables 0.6.15